+++ dummyfilescreator/qt_about_window.py
